edit import ptk

+++ app/Http/Controllers/Backend/PtkController.php
@@ -2,11 +2,13 @@
 
 namespace App\Http\Controllers\Backend;
 
-use App\Imports\ImportPtk;
+use App\Models\Sekolah;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
+use App\Imports\ImportCollectionPtk;
 use Illuminate\Support\Facades\File;
 use Maatwebsite\Excel\Facades\Excel;
+use Illuminate\Support\Facades\Storage;
 
 class PtkController extends Controller
 {
@@ -19,7 +21,7 @@ class PtkController extends Controller
     */
     public function __construct()
     {
-        $this->uploadPath = public_path(config('cms.image.directoryLogo'));
+        $this->uploadPath = public_path(config('cms.image.directoryPtk'));
     }
 
 
@@ -36,6 +38,7 @@ class PtkController extends Controller
     */
     public function index()
     {
+        $this->cleanupload();
         return view('backend.ptk.index', [
             'title' => 'Daftar PTK'
         ]);
@@ -48,7 +51,9 @@ class PtkController extends Controller
     */
     public function create()
     {
+        $this->cleanupload();
         return view('backend.ptk.create', [
+            'sekolah' => Sekolah::all(),
             'title' => 'Daftar PTK'
         ]);
 
@@ -60,7 +65,8 @@ class PtkController extends Controller
             'importfile' => 'required|mimes:xls,xlsx,csv'
         ]);
 
-        $file = $request->file('importfile');
+        $sekolahId = $request->input('sekolah_id');
+        $file      = $request->file('importfile');
 
         $nama_file = $file->hashName();
 
@@ -70,12 +76,29 @@ class PtkController extends Controller
 
 
         // import data
-        $import = Excel::import(new ImportPtk(), ('uploads/files/excel/'.$nama_file));
+        // $import = Excel::import(new ImportPtk(), ('uploads/files/excel/'.$nama_file));
+        $import = new ImportCollectionPtk(
+            $sekolahId,
+        );
 
+        $import->import('uploads/files/excel/'.$nama_file);
 
+        if ($import->failures()->isNotEmpty()) {
+            return back()->withFailures($import->failures());
+        }
         //remove file import from server
         File::delete('uploads/files/excel/'.$nama_file);
 
         return redirect()->route('backend.ptk.index')->with('success', 'Data PTK berhasil diimport!');
+    }
+
+    public function cleanupload()
+    {
+        $tempImages = Storage::files('files/excel');
+
+        foreach ($tempImages as $file) {
+            Storage::delete($file);
+        }
+
     }
 }
